requirements.txt
pypresence==4.2.1
websocket_client==1.4.2
